+++ app/shop/src/main/java/com/shoppingmall/vo/Cart.java
@@ -9,10 +9,10 @@ import java.time.LocalDateTime;
 @Data
 @NoArgsConstructor
 public class Cart {
-    private Long cartId;
-    private Long memberId;
-    private Long productId;
-    private Long amount;
+    private Integer cartId;
+    private Integer memberId;
+    private Integer productId;
+    private Integer amount;
     private LocalDateTime createDate;
     private LocalDateTime updateDate;
     private Integer totalPricePerCartItem; // 장바구니 상품별 총 합계
@@ -20,7 +20,7 @@ public class Cart {
     private Member member;
 
     @Builder
-    public Cart(Long cartId, Long memberId, Long productId, Long amount, LocalDateTime createDate, LocalDateTime updateDate, Integer totalPricePerCartItem, Product product, Member member) {
+    public Cart(Integer cartId, Integer memberId, Integer productId, Integer amount, LocalDateTime createDate, LocalDateTime updateDate, Integer totalPricePerCartItem, Product product, Member member) {
         this.cartId = cartId;
         this.memberId = memberId;
         this.productId = productId;
